+++ lib/searchkick/bulk_reindex_job.rb
@@ -2,7 +2,7 @@ module Searchkick
   class BulkReindexJob < Searchkick.parent_job.constantize
     queue_as { Searchkick.queue_name }
 
-    def perform(class_name:, record_ids: nil, index_name: nil, method_name: nil, batch_id: nil, min_id: nil, max_id: nil, allow_missing: nil)
+    def perform(class_name:, record_ids: nil, index_name: nil, method_name: nil, batch_id: nil, min_id: nil, max_id: nil, ignore_missing: nil)
       model = Searchkick.load_model(class_name)
       index = model.searchkick_index(name: index_name)
 
@@ -12,7 +12,7 @@ module Searchkick
       relation = Searchkick.load_records(relation, record_ids)
       relation = relation.search_import if relation.respond_to?(:search_import)
 
-      RecordIndexer.new(index).reindex(relation, mode: :inline, method_name: method_name, allow_missing: allow_missing, full: false)
+      RecordIndexer.new(index).reindex(relation, mode: :inline, method_name: method_name, ignore_missing: ignore_missing, full: false)
       RelationIndexer.new(index).batch_completed(batch_id) if batch_id
     end
   end