+++ lib/searchkick/indexer.rb
@@ -24,7 +24,7 @@ module Searchkick
         # note: delete does not set error when item not found
         first_with_error = response["items"].map do |item|
           (item["index"] || item["delete"] || item["update"])
-        end.find.with_index { |item, i| item["error"] && !allow_missing?(items[i], item["error"]) }
+        end.find.with_index { |item, i| item["error"] && !ignore_missing?(items[i], item["error"]) }
         if first_with_error
           raise ImportError, "#{first_with_error["error"]} on item with id '#{first_with_error["_id"]}'"
         end
@@ -36,8 +36,8 @@ module Searchkick
 
     private
 
-    def allow_missing?(item, error)
-      error["type"] == "document_missing_exception" && item.instance_variable_defined?(:@allow_missing)
+    def ignore_missing?(item, error)
+      error["type"] == "document_missing_exception" && item.instance_variable_defined?(:@ignore_missing)
     end
   end
 end