+++ test/partial_reindex_test.rb
@@ -54,15 +54,15 @@ class PartialReindexTest < Minitest::Test
     assert_match "document missing", error.message
   end
 
-  def test_record_allow_missing_inline
+  def test_record_ignore_missing_inline
     store [{name: "Hi", color: "Blue"}]
 
     product = Product.first
     Product.searchkick_index.remove(product)
 
-    product.reindex(:search_name, allow_missing: true)
+    product.reindex(:search_name, ignore_missing: true)
     Searchkick.callbacks(:bulk) do
-      product.reindex(:search_name, allow_missing: true)
+      product.reindex(:search_name, ignore_missing: true)
     end
   end
 
@@ -80,14 +80,14 @@ class PartialReindexTest < Minitest::Test
     end
   end
 
-  def test_record_allow_missing_async
+  def test_record_ignore_missing_async
     store [{name: "Hi", color: "Blue"}]
 
     product = Product.first
     Product.searchkick_index.remove(product)
 
     perform_enqueued_jobs do
-      product.reindex(:search_name, mode: :async, allow_missing: true)
+      product.reindex(:search_name, mode: :async, ignore_missing: true)
     end
   end
 
@@ -146,13 +146,13 @@ class PartialReindexTest < Minitest::Test
     assert_match "document missing", error.message
   end
 
-  def test_relation_allow_missing_inline
+  def test_relation_ignore_missing_inline
     store [{name: "Hi", color: "Blue"}]
 
     product = Product.first
     Product.searchkick_index.remove(product)
 
-    Product.where(id: product.id).reindex(:search_name, allow_missing: true)
+    Product.where(id: product.id).reindex(:search_name, ignore_missing: true)
   end
 
   def test_relation_missing_async
@@ -169,14 +169,14 @@ class PartialReindexTest < Minitest::Test
     end
   end
 
-  def test_relation_allow_missing_async
+  def test_relation_ignore_missing_async
     store [{name: "Hi", color: "Blue"}]
 
     product = Product.first
     Product.searchkick_index.remove(product)
 
     perform_enqueued_jobs do
-      Product.where(id: product.id).reindex(:search_name, mode: :async, allow_missing: true)
+      Product.where(id: product.id).reindex(:search_name, mode: :async, ignore_missing: true)
     end
   end
 end